Add Clash.Normalize.Transformations.XOptimize
All transformations concerning X optimization now live in a separate
module. This is part of work to make the normalization modules
quicker to compile.
